requirements.txt
# Pinned dependencies for Bloomberg Elite V3 (Streamlit Cloud Stable)
streamlit==1.35.0
yfinance==0.2.52
pandas==2.2.2
numpy==1.26.4
plotly==5.22.0
google-generativeai==0.3.0
requests==2.32.3
cot_reports==0.1.3
beautifulsoup4==4.12.3
feedparser==6.0.11
altair==5.3.0
matplotlib==3.9.0
scipy==1.13.1
protobuf==4.25.3
python-dateutil==2.9.0
Pillow==10.3.0
